+++ janitor/functions/conditional_join.py
@@ -27,6 +27,7 @@ from ._conditional_join import (
     _get_indices_equi,
     _get_indices_non_equi,
     _get_indices_single_join,
+    _get_join_aggs,
     _not_equal_indices,
 )
 from ._conditional_join._helpers import (
@@ -41,7 +42,7 @@ from ._conditional_join._helpers import (
 def conditional_join(
     df: pd.DataFrame,
     right: pd.DataFrame | pd.Series,
-    *conditions: Any,
+    *conditions: tuple,
     how: Literal["inner", "left", "right", "outer"] = "inner",
     df_columns: Optional[Any] = slice(None),
     right_columns: Optional[Any] = slice(None),
@@ -49,6 +50,8 @@ def conditional_join(
     use_numba: bool = False,
     indicator: Optional[bool | str] = False,
     force: bool = False,
+    join_algorithm: str = "default",
+    include_join_positions: bool = False,
 ) -> pd.DataFrame:
     """The conditional_join function operates similarly to `pd.merge`,
     but supports joins on inequality operators,
@@ -107,6 +110,12 @@ def conditional_join(
     If the columns from `df` and `right` have nothing in common,
     a single index column is returned; else, a MultiIndex column
     is returned.
+
+    If `include_join_positions` is `True`, the index of the returned dataframe
+    will be a MultiIndex; the first level points to the original positions in `df`,
+    while the second level points to the original positions in `right`.
+
+
 
     Examples:
         >>> import pandas as pd
@@ -250,6 +259,9 @@ def conditional_join(
             - `col` class deprecated.
         - 0.33.0
             - `use_numba` deprecated.
+        - 0.34.0
+            - Added `include_join_positions` parameter.
+            - Added `join_algorithm` parameter.
 
     Args:
         df: A pandas DataFrame.
@@ -282,6 +294,11 @@ def conditional_join(
             only appears in the right DataFrame, and `both` if the observation’s
             merge key is found in both DataFrames.
         force: If `True`, force the non-equi join conditions to execute before the equi join.
+        join_algorithm: Determines what algorithm to use for multiple non-equi joins.
+            Currently limited to `default` and `regions`.
+        include_join_positions: Determines if the join positions of the left and right DataFrame
+            should be included as an index of the final dataframe.
+
 
 
     Returns:
@@ -299,6 +316,12 @@ def conditional_join(
         use_numba=use_numba,
         indicator=indicator,
         force=force,
+        aggfunc=None,
+        include_join_positions=include_join_positions,
+        return_building_blocks=False,
+        reverse=False,
+        return_matching_indices=False,
+        join_algorithm=join_algorithm,
     )
 
 
@@ -328,6 +351,11 @@ def _conditional_join_preliminary_checks(
     indicator: bool | str,
     force: bool,
     return_matching_indices: bool = False,
+    aggfunc: list[tuple] = None,
+    include_join_positions: bool = False,
+    return_building_blocks: bool = False,
+    reverse: bool = False,
+    join_algorithm: str = "default",
 ) -> tuple:
     """
     Preliminary checks for conditional_join are conducted here.
@@ -356,14 +384,6 @@ def _conditional_join_preliminary_checks(
             f"from the right dataframe is {right.columns.nlevels}."
         )
 
-    # Check MultiIndex dictionary renaming before column existence checks
-    if (df.columns.nlevels > 1) and (
-        isinstance(df_columns, dict) or isinstance(right_columns, dict)
-    ):
-        raise ValueError(
-            "Column renaming with a dictionary is not supported for MultiIndex columns."
-        )
-
     if not conditions:
         raise ValueError("Kindly provide at least one join condition.")
 
@@ -384,12 +404,6 @@ def _conditional_join_preliminary_checks(
         check_column(right, [right_on])
         _check_operator(op)
 
-    if (
-        all((op == _JoinOperator.STRICTLY_EQUAL.value for *_, op in conditions))
-        and not return_matching_indices
-    ):
-        raise ValueError("Equality only joins are not supported.")
-
     check("how", how, [str])
 
     if how not in {"inner", "left", "right", "outer"}:
@@ -400,7 +414,7 @@ def _conditional_join_preliminary_checks(
     if keep not in {"all", "first", "last"}:
         raise ValueError("'keep' should be one of 'all', 'first', 'last'.")
 
-    # deprecate in a future version
+    # TODO: deprecate in a future version
     check("use_numba", use_numba, [bool])
 
     if use_numba:
@@ -409,6 +423,67 @@ def _conditional_join_preliminary_checks(
     check("indicator", indicator, [bool, str])
 
     check("force", force, [bool])
+
+    check("reverse", reverse, [bool])
+
+    if aggfunc is not None:
+        check("aggfunc", aggfunc, [list])
+        if all((op == _JoinOperator.NOT_EQUAL.value for *_, op in conditions)):
+            raise NotImplementedError(
+                "aggfunc is not supported when all the join operators are !="
+            )
+        if reverse:
+            cols = df.columns
+            frame = df
+            replacement = "left"
+        else:
+            cols = right.columns
+            frame = right
+            replacement = "right"
+        for entry in aggfunc:
+            check("entry in aggfunc", entry, [tuple])
+            if len(entry) != 2:
+                raise ValueError(
+                    "The tuple in an aggfunc should be 2 elements; "
+                    "The first element in the tuple should be a column name "
+                    f"in the {replacement} dataframe, while the second element "
+                    "in the tuple should be a supported aggregation function"
+                )
+        aggs = {"sum", "min", "max", "size", "prod"}
+        for column_name, agg in aggfunc:
+            if column_name not in cols:
+                raise KeyError(
+                    f"{column_name} in aggfunc does not "
+                    f"exist in the {replacement} dataframe"
+                )
+            if agg not in aggs:
+                raise ValueError(
+                    f"The aggregation function for {column_name} "
+                    f"should be one of {','.join(aggs)}; "
+                    f"instead got {agg}"
+                )
+            if (agg in {"sum", "prod"}) and not pd.api.types.is_numeric_dtype(
+                frame[column_name]
+            ):
+                raise ValueError(f"{agg} is supported only for numeric columns")
+    if all((op == _JoinOperator.STRICTLY_EQUAL.value for *_, op in conditions)):
+        if not (return_matching_indices or aggfunc):
+            raise ValueError("Equality only joins are not supported.")
+        if return_matching_indices and use_numba:
+            raise ValueError(
+                "Equality only joins are supported only if use_numba is False."
+            )
+
+    check("include_join_positions", include_join_positions, [bool])
+    if include_join_positions and (how != "inner"):
+        raise ValueError("include_join_positions is valid only if `how='inner'`")
+    check("return_building_blocks", return_building_blocks, [bool])
+    check("join_algorithm", join_algorithm, [str])
+    if join_algorithm not in {"default", "regions"}:
+        raise ValueError(
+            f"join_algorithm should be either default or regions, "
+            f"instead got {join_algorithm}"
+        )
 
     return (
         df,
@@ -421,6 +496,11 @@ def _conditional_join_preliminary_checks(
         use_numba,
         indicator,
         force,
+        aggfunc,
+        include_join_positions,
+        return_building_blocks,
+        reverse,
+        join_algorithm,
     )
 
 
@@ -472,6 +552,11 @@ def _conditional_join_compute(
     indicator: bool | str,
     force: bool,
     return_matching_indices: bool = False,
+    aggfunc: list[tuple] = None,
+    include_join_positions: bool = False,
+    return_building_blocks: bool = False,
+    reverse: bool = False,
+    join_algorithm: str = "default",
 ) -> pd.DataFrame:
     """
     This is where the actual computation
@@ -488,6 +573,11 @@ def _conditional_join_compute(
         use_numba,
         indicator,
         force,
+        aggfunc,
+        include_join_positions,
+        return_building_blocks,
+        reverse,
+        join_algorithm,
     ) = _conditional_join_preliminary_checks(
         df=df,
         right=right,
@@ -500,6 +590,11 @@ def _conditional_join_compute(
         indicator=indicator,
         force=force,
         return_matching_indices=return_matching_indices,
+        aggfunc=aggfunc,
+        include_join_positions=include_join_positions,
+        return_building_blocks=return_building_blocks,
+        reverse=reverse,
+        join_algorithm=join_algorithm,
     )
     eq_check = False
     le_lt_check = False
@@ -518,47 +613,65 @@ def _conditional_join_compute(
     df.index = range(len(df))
     right.index = range(len(right))
     if eq_check:
-        result = _multiple_conditional_join_eq(
+        indices = _multiple_conditional_join_eq(
             df=df,
             right=right,
             conditions=conditions,
             keep=keep,
             use_numba=use_numba,
             force=force,
+            return_matching_indices=return_building_blocks or aggfunc,
+            join_algorithm=join_algorithm,
         )
     elif (len(conditions) > 1) & le_lt_check:
-        result = _multiple_conditional_join_le_lt(
+        indices = _multiple_conditional_join_le_lt(
             df=df,
             right=right,
             conditions=conditions,
             keep=keep,
             use_numba=use_numba,
+            return_matching_indices=return_building_blocks or aggfunc,
+            join_algorithm=join_algorithm,
         )
     elif len(conditions) > 1:
-        result = _multiple_conditional_join_ne(
+        indices = _multiple_conditional_join_ne(
             df=df,
             right=right,
             conditions=conditions,
             keep=keep,
         )
     else:
-        result = _get_indices_single_join._single_join(
+        indices = _get_indices_single_join._single_join(
             df=df,
             right=right,
             condition=conditions[0],
             keep=keep,
+            return_matching_indices=return_building_blocks or aggfunc,
+        )
+    if aggfunc and reverse:
+        return _get_join_aggs._agg_join_left(
+            df=df,
+            aggfunc=aggfunc,
+            indices=indices,
+        )
+    if aggfunc:
+        return _get_join_aggs._agg_join_right(
+            right=right,
+            aggfunc=aggfunc,
+            indices=indices,
         )
     if return_matching_indices:
-        return result
+        return indices
     return _create_frame(
         df=df,
         right=right,
-        left_index=result["left_index"],
-        right_index=result["right_index"],
+        left_index=indices["left_index"],
+        right_index=indices["right_index"],
         how=how,
         df_columns=df_columns,
         right_columns=right_columns,
         indicator=indicator,
+        include_join_positions=include_join_positions,
     )
 
 
@@ -657,6 +770,8 @@ def _multiple_conditional_join_eq(
     keep: str,
     use_numba: bool,
     force: bool,
+    return_matching_indices: bool,
+    join_algorithm: str,
 ) -> tuple:
     """
     Get indices for multiple conditions,
@@ -672,6 +787,8 @@ def _multiple_conditional_join_eq(
             conditions=conditions,
             keep=keep,
             use_numba=use_numba,
+            return_matching_indices=return_matching_indices,
+            join_algorithm=join_algorithm,
         )
     # deprecated - no longer maintained
     if use_numba:
@@ -787,6 +904,7 @@ def _multiple_conditional_join_eq(
         right=right,
         conditions=conditions,
         keep=keep,
+        return_matching_indices=return_matching_indices,
     )
 
 
@@ -796,6 +914,8 @@ def _multiple_conditional_join_le_lt(
     conditions: list,
     keep: str,
     use_numba: bool,
+    return_matching_indices: bool,
+    join_algorithm: str,
 ) -> tuple:
     """
     Get indices for multiple conditions,
@@ -857,6 +977,8 @@ def _multiple_conditional_join_le_lt(
         right=right,
         conditions=conditions,
         keep=keep,
+        return_matching_indices=return_matching_indices,
+        join_algorithm=join_algorithm,
     )
 
 
@@ -888,6 +1010,7 @@ def _create_frame(
     df_columns: Any,
     right_columns: Any,
     indicator: bool | str,
+    include_join_positions: bool,
 ) -> pd.DataFrame:
     """
     Create final dataframe
@@ -958,6 +1081,7 @@ def _create_frame(
         left_index: np.ndarray,
         right_index: np.ndarray,
         indicator: bool | str,
+        include_join_positions: bool = False,
     ) -> pd.DataFrame:
         """Computes an inner joined DataFrame.
 
@@ -967,7 +1091,9 @@ def _create_frame(
             left_index: indices from df for rows that match right.
             right_index: indices from right for rows that match df.
             indicator: Indicator column name or True for default name "_merge".
-
+            include_join_positions: Determines if the join positions of the left
+                and right DataFrame should be included as an index
+                of the final dataframe.
         Returns:
             An inner joined DataFrame.
         """
@@ -984,6 +1110,9 @@ def _create_frame(
                 columns=df.columns.union(right.columns),
             )
             dictionary[indicator] = arr
+        if include_join_positions:
+            index = pd.MultiIndex.from_arrays([left_index, right_index])
+            return pd.DataFrame(dictionary, copy=False, index=index)
         return pd.DataFrame(dictionary, copy=False)
 
     if how == "inner":
@@ -993,6 +1122,7 @@ def _create_frame(
             left_index=left_index,
             right_index=right_index,
             indicator=indicator,
+            include_join_positions=include_join_positions,
         )
     if how == "left":
         indexer = pd.unique(left_index)
@@ -1168,10 +1298,12 @@ def _create_frame(
 def get_join_indices(
     df: pd.DataFrame,
     right: pd.DataFrame | pd.Series,
-    conditions: list[tuple[str]],
+    *conditions: tuple,
     keep: Literal["first", "last", "all"] = "all",
     use_numba: bool = False,
     force: bool = False,
+    return_building_blocks: bool = False,
+    join_algorithm: str = "default",
 ) -> dict:
     """Convenience function to return the matching indices from an inner join.
 
@@ -1186,11 +1318,14 @@ def get_join_indices(
             - return indices as a dictionary.
         - 0.33.0
             - `use_numba` deprecated.
+        - 0.34.0
+            - Added experimental `return_building_blocks` parameter.
+            - Add join_algorithm parameter.
 
     Args:
         df: A pandas DataFrame.
         right: Named Series or DataFrame to join to.
-        conditions: List of arguments of tuple(s) of the form
+        conditions: Variable arguments of tuple(s) of the form
             `(left_on, right_on, op)`, where `left_on` is the column
             label from `df`, `right_on` is the column label from `right`,
             while `op` is the operator.
@@ -1203,6 +1338,11 @@ def get_join_indices(
         keep: Choose whether to return the first match, last match or all matches.
         force: If `True`, force the non-equi join conditions
             to execute before the equi join.
+        return_building_blocks: Return a possibly more extensive dictionary,
+            containing data that will be used to build the indices.
+            !!! warning "This feature is experimental and may change without warning."
+        join_algorithm: Determines what algorithm to use for multiple non-equi joins.
+            Currently limited to `default` and `regions`.
 
     Returns:
         A dictionary of indices for the rows in the dataframes that match.
@@ -1219,6 +1359,138 @@ def get_join_indices(
         indicator=False,
         force=force,
         return_matching_indices=True,
+        aggfunc=None,
+        include_join_positions=False,
+        return_building_blocks=return_building_blocks,
+        reverse=False,
+        join_algorithm=join_algorithm,
+    )
+
+
+@pf.register_dataframe_method
+def join_agg(
+    df: pd.DataFrame,
+    right: pd.DataFrame | pd.Series,
+    *conditions,
+    aggfunc: list[tuple],
+    force: bool = False,
+    reverse: bool = False,
+    join_algorithm: str = "default",
+) -> pd.DataFrame:
+    """
+    Compute an aggregation after the successful execution of a join;
+    the aggregaton is computed on the right dataframe
+    for each row of the left DataFrame (that has a match)
+    based on the join keys.
+
+    If `reverse=True`, the aggregaton is computed
+    on the left dataframe for each row of the right DataFrame
+    (that has a match) based on the join keys.
+
+    Supported aggregation functions are
+    `sum`, `prod`, `size`, `min`, `max`.
+
+    This is limited to an inner join.
+
+    The index of the returned dataframe represent the positions
+    of the rows from the left dataframe that have matches
+    in the right dataframe.
+
+    If `reverse=True`, the index of the returned dataframe
+    represent the positions of the rows from the right dataframe
+    that have matches in the left dataframe.
+
+    !!! info "New in version 0.34.0"
+
+    Examples:
+        >>> import pandas as pd
+        >>> import janitor
+        >>> df1 = pd.DataFrame(
+        ...     {"id": [1, 1, 1, 2, 2, 3], "value_1": [2, 5, 7, 1, 3, 4]}
+        ... )
+        >>> df2 = pd.DataFrame(
+        ...     {
+        ...         "id": [1, 1, 1, 1, 2, 2, 2, 3],
+        ...         "value_2A": [0, 3, 7, 12, 0, 2, 3, 1],
+        ...         "value_2B": [1, 5, 9, 15, 1, 4, 6, 3],
+        ...     }
+        ... )
+        >>> df1
+            id  value_1
+         0   1        2
+         1   1        5
+         2   1        7
+         3   2        1
+         4   2        3
+         5   3        4
+        >>> df2
+            id  value_2A  value_2B
+         0   1         0         1
+         1   1         3         5
+         2   1         7         9
+         3   1        12        15
+         4   2         0         1
+         5   2         2         4
+         6   2         3         6
+         7   3         1         3
+        >>> (
+        ...     df1.join_agg(
+        ...         df2,
+        ...         ("id", "id", "=="),
+        ...         ("value_1", "value_2A", ">="),
+        ...         ("value_1", "value_2B", "<="),
+        ...         aggfunc=[("value_2A", "sum"), ("value_2B", "min"), ("id", "size")],
+        ...     )
+        ... )
+          value_2A value_2B   id
+               sum      min size
+        1        3        5    1
+        2        7        9    1
+        3        0        1    1
+        4        5        4    2
+
+    Args:
+        df: A pandas DataFrame.
+        right: Named Series or DataFrame to join to.
+        conditions: Variable arguments of tuple(s) of the form
+            `(left_on, right_on, op)`, where `left_on` is the column
+            label from `df`, `right_on` is the column label from `right`,
+            while `op` is the operator.
+            The operator can be any of
+            `==`, `!=`, `<=`, `<`, `>=`, `>`. For multiple conditions,
+            the and(`&`) operator is used to combine the results
+            of the individual conditions.
+        force: If `True`, force the non-equi join conditions
+            to execute before the equi join.
+        aggfunc: Compute aggregates on the right dataframe
+            for each row of the left DataFrame (that has a match)
+            based on the join keys.
+            Supported aggregation functions are
+            `sum`, `size`, `min`, `max`, `prod`.
+        reverse: If `True`, compute the aggregation on the columns
+            of the left dataframe; if `False`, which is the default,
+            compute the aggregation on the columns of the right dataframe.
+        join_algorithm: Determines what algorithm to use for multiple non-equi joins.
+            Currently limited to `default` and `regions`.
+
+    Returns:
+        A pandas DataFrame.
+    """
+    return _conditional_join_compute(
+        df=df,
+        right=right,
+        conditions=conditions,
+        how="inner",
+        df_columns=None,
+        right_columns=None,
+        keep="all",
+        use_numba=False,
+        indicator=False,
+        force=force,
+        return_matching_indices=False,
+        aggfunc=aggfunc,
+        reverse=reverse,
+        join_algorithm=join_algorithm,
     )